requirements.txt
Flask
gunicorn
elasticsearch
pandas
python-dotenv
requests

